requirements.txt
azure-identity
azure-mgmt-resourcegraph
requests
